# api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: EcoSignal API Gateway
  description: >
    Public API for EcoSignal — connects Italian citizens' daily habits
    to real local environmental data, generating personalized climate
    narratives and actionable CO₂-reduction suggestions.
  version: 1.0.0
  contact:
    name: EcoSignal Team
servers:
  - url: http://localhost:8000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransportMode:
      type: string
      enum: [car, transit, bike, walk, mixed]
    DietType:
      type: string
      enum: [meat_daily, meat_weekly, vegetarian, vegan]
    HomeType:
      type: string
      enum: [apartment, house]
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 3600

    # ── User Profile / Onboarding ─────────────────────────
    OnboardingRequest:
      type: object
      required: [zip_code, transport_mode, diet_type, home_type, home_size_sqm]
      properties:
        zip_code:
          type: string
          pattern: '^\d{5}$'
          description: Italian CAP (5 digits)
          example: "20121"
        transport_mode:
          $ref: '#/components/schemas/TransportMode'
        diet_type:
          $ref: '#/components/schemas/DietType'
        home_type:
          $ref: '#/components/schemas/HomeType'
        home_size_sqm:
          type: integer
          minimum: 10
          maximum: 500
          example: 75
    ProfileUpdateRequest:
      type: object
      minProperties: 1
      properties:
        zip_code:
          type: string
          pattern: '^\d{5}$'
        transport_mode:
          $ref: '#/components/schemas/TransportMode'
        diet_type:
          $ref: '#/components/schemas/DietType'
        home_type:
          $ref: '#/components/schemas/HomeType'
        home_size_sqm:
          type: integer
          minimum: 10
          maximum: 500
    HabitProfile:
      type: object
      required:
        - user_id
        - zip_code
        - transport_mode
        - diet_type
        - home_type
        - home_size_sqm
        - created_at
        - updated_at
      properties:
        user_id:
          type: string
          format: uuid
        zip_code:
          type: string
          pattern: '^\d{5}$'
        transport_mode:
          $ref: '#/components/schemas/TransportMode'
        diet_type:
          $ref: '#/components/schemas/DietType'
        home_type:
          $ref: '#/components/schemas/HomeType'
        home_size_sqm:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Footprint ─────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Invalid or expired token
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created, tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Validation error (e.g. weak password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Authenticate and obtain tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Refresh an expired access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Profile ───────────────────────────────────────────
  /api/v1/onboarding:
    post:
      tags: [profile]
      operationId: submitOnboarding
      summary: Submit initial habit profile during onboarding
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitProfile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Profile already exists for this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/profile:
    get:
      tags: [profile]
      operationId: getProfile
      summary: Get current user's habit profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found — onboarding not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [profile]
      operationId: updateProfile
      summary: Update one or more habit fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitProfile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Footprint ─────────────────────────────────────────
...